@@ -1,22 +1,24 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http'
+import { NgModule } from '@angular/core';
+import { HttpClientModule } from '@angular/common/http';
+import { MatListModule } from '@angular/material/list';
 
 const modules = [
   CommonModule,
   FormsModule,
   HttpClientModule,
+  MatListModule,
+  
 ]
 
 @NgModule({
   declarations: [
+
   ],
-  imports: [
-    ...modules
-  ],
-  exports: [
-    ...modules
-  ]
+  imports: [...modules],
+  exports: [...modules],
 })
-export class CommonsModule { }
+export class CommonsModule {
+
+}
