@@ -10,7 +10,7 @@ describe('workspace-project App', () => {
 
   it('should display welcome message', () => {
     page.navigateTo();
-    expect(page.getTitleText()).toEqual('curso-angular-feb12 app is running!');
+    expect(page.getTitleText()).toEqual('Angular-Project app is running!');
   });
 
   afterEach(async () => {
