@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 import { PlaceholderService } from './services/placeholder.service';
-import { Usuario } from './models/user.model';
 
 @Component({
   selector: 'app-root',
@@ -8,42 +7,16 @@ import { Usuario } from './models/user.model';
   styleUrls: ['./app.component.css']
 })
 export class AppComponent implements OnInit {
-
-  title = 'curso-angular-feb12';
-  nombreDeBoton = "haz click";
-  valorDeInput = "Hola!";
-  list: Usuario[];
-
+  users = [];
+ title ="hola hans"
   constructor(private phService: PlaceholderService) { }
-
-  ngOnInit() {
-    this.phService.getUsersAlt().subscribe(users => {
-      console.log(users)
-      this.mostrarUsuarios(users)
-    });
-    console.log("hola");
-
-
-
+  
+  async ngOnInit() {
+    this.phService.getUsersAlt().subscribe(users => this.users = users);
   }
 
-
-  mostrarUsuarios(e: Usuario[]) {
-    console.log(e)
-    this.list = e
+  deleteFirst ():void {
+    this.users.shift();
   }
-
-  cambiarTitulo(): void {
-    this.title = "Otro titulo";
-    this.nombreDeBoton = "click realizado";
-    this.valorDeInput = "Nuevo valor";
-  }
-
-  vaciarLista(): void {
-    this.list = []
-  }
-
-  imprimirEnConsola(): void {
-    console.log(this.valorDeInput);
-  }
+  
 }
