@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Usuario } from '../models/user.model';
+import { User } from '../models/user.model';
 import { Observable } from 'rxjs';
 
 @Injectable({
@@ -9,14 +9,18 @@ import { Observable } from 'rxjs';
 export class PlaceholderService {
 
   constructor(private httpClient: HttpClient) { }
+  async getUsers(): Promise <User[]> {
+    let users = await this.httpClient.get<User[]>(
+      'https://jsonplaceholder.typicode.com/users',
+    ).toPromise();
 
-  async getUsers(): Promise<Usuario[]> {
-    let usuarios = await this.httpClient.get<Usuario[]>("https://jsonplaceholder.typicode.com/users").toPromise()
-    return usuarios
+    return users;
   }
+  getUsersAlt(): Observable <User[]> {
+    let observableUsers = this.httpClient.get<User[]>(
+      'https://jsonplaceholder.typicode.com/users',
+    );
 
-  getUsersAlt(): Observable<Usuario[]> {
-    return this.httpClient.get<Usuario[]>("https://jsonplaceholder.typicode.com/users")
+    return observableUsers;
   }
-
 }
